Refactored Subscription components and established connection to backend
