+++ src/pages/Billing/ProductDescription.tsx
@@ -1,0 +1,14 @@
+export default function ProductDescription() {
+  return (
+    <div>
+      <h1 className="my-4">Everything from Free, plus:</h1>
+      <ul>
+        <li className="mb-2">Unlimited projects</li>
+        <li className="mb-2">Unlimited storage</li>
+        <li className="mb-2">Advanced analytics</li>
+        <li className="mb-2">Custom permissions</li>
+        <li className="mb-2">Advanced integrations</li>
+      </ul>
+    </div>
+  );
+}
